# openapi.yaml
openapi: 3.1.0
info:
  title: Nagarik App Platform API
  version: 1.0.0
  summary: Consolidated OpenAPI specification for the Nagarik App service platform.
  description: |
    This specification defines versioned REST APIs for the Nagarik App platform.
    It consolidates the public service surfaces for identity and consent, document intake,
    OCR extraction, tax guidance, application tracking, service orchestration, notifications,
    reference policy, and audit records.
servers:
  - url: https://api.nagarik.gov.np/v1
    description: Primary government production API
  - url: https://staging-api.nagarik.gov.np/v1
    description: Staging API
security:
  - bearerAuth: []
...
paths:
...
  /applications/{trackingReference}:
    get:
      tags: [Application Tracking]
      operationId: getApplication
      summary: Get a tracked application.
      x-authorization-requirements:
        - Valid bearer token
        - Citizen access or authorized support access
      parameters:
        - $ref: '#/components/parameters/TrackingReference'
      responses:
        '200':
          description: Application tracking detail.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /applications:
    get:
      tags: [Application Tracking]
      operationId: listApplications
      summary: List tracked applications.
      x-authorization-requirements:
        - Valid bearer token
        - Citizen access or authorized support access
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
        - name: service_code
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Paginated application list.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Maximum number of items to return.
    Offset:
      name: offset
      in: query
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
      description: Number of items to skip.
...
    TrackingReference:
      name: trackingReference
      in: path
      required: true
      schema:
        type: string
...
  responses:
...
    Unauthorized:
      description: Authentication required or failed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: The authenticated principal is not authorized.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: The requested resource was not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [error, message, trace_id]
      properties:
        error:
          type: string
        message:
          type: string
        trace_id:
          type: string
        details:
          type: array
          items:
            $ref: '#/components/schemas/ErrorDetail'
    ErrorDetail:
      type: object
      required: [field, issue]
      properties:
        field:
          type: string
        issue:
          type: string
...
    PaginationMeta:
      type: object
      required: [limit, offset, total]
      properties:
        limit:
          type: integer
        offset:
          type: integer
        total:
          type: integer
...
    ApplicationResponse:
      type: object
      required: [tracking_reference, current_status_code, submitted_at]
      properties:
        id:
          type: string
          format: uuid
        tracking_reference:
          type: string
        external_application_ref:
          type: string
          nullable: true
        current_status_code:
          type: string
        submitted_at:
          type: string
          format: date-time
        last_status_at:
          type: string
          format: date-time
        is_sensitive:
          type: boolean
        service_definition_id:
          type: string
          format: uuid
...
    ApplicationListResponse:
      type: object
      required: [items, page]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ApplicationResponse'
        page:
          $ref: '#/components/schemas/PaginationMeta'
...